use ai network blockchain for storing root of the dag.
